mechanicus_laser_cad/requirements.txt
altgraph==0.17.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fonttools==4.60.0
future==1.0.0
greenlet==3.2.4
imageio==2.37.0
install-playwright==0.1.1
iso8601==2.1.0
keyboard==0.13.5
kiwisolver==1.4.9
lazy_loader==0.4
matplotlib==3.10.6
networkx==3.5
numpy==1.26.4
opencv-python==4.12.0.88
packaging==24.2
pefile==2023.2.7
pillow==10.4.0
playwright==1.55.0
pycairo==1.28.0
pyee==13.0.0
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pyparsing==3.2.5
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
PyYAML==6.0.2
scikit-image==0.25.2
scipy==1.16.2
serial==0.0.97
setuptools==75.8.0
six==1.17.0
svg.path==7.0
svgpathtools==1.7.1
svgtrace==2024
svgwrite==1.4.3
tifffile==2025.9.20
tk==0.1.0
tkcolorpicker==2.1.3
typing_extensions==4.15.0
